requirements.txt
pyserial==3.4
python-dotenv==0.14.0
pytz==2020.5
requests==2.24.0
